Shoppingmoa/requirements.txt
Appium-Python-Client==1.1.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
cycler==0.10.0
et-xmlfile==1.1.0
Flask==2.0.1
html-testRunner==1.2.1
idna==2.10
imageio==2.9.0
imgkit==1.2.2
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.3
mysql-connector-python==8.0.26
networkx==2.6.2
numpy==1.20.2
opencv-python==4.5.3.56
openpyxl==3.0.7
pandas==1.2.4
Pillow==8.2.0
protobuf==3.17.3
pycparser==2.20
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
requests==2.25.1
scikit-image==0.18.3
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.1
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
threadpoolctl==2.2.0
tifffile==2021.8.30
urllib3==1.26.4
Werkzeug==2.0.1
